requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pandas
python-dotenv
google-generativeai
openpyxl
python-multipart